requirements.txt
cx-Freeze==5.1.1
idna==2.8
lark-parser==0.9.0
